requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0
torchmetrics>=1.0.0
Pillow>=9.0.0
PyYAML>=6.0
numpy>=1.21.0 